Delete report views
