+++ crates/api/api/Cargo.toml
@@ -28,7 +28,6 @@ lemmy_db_views_post = { workspace = true, features = ["full"] }
 lemmy_db_views_vote = { workspace = true, features = ["full"] }
 lemmy_db_views_local_user = { workspace = true, features = ["full"] }
 lemmy_db_views_person = { workspace = true, features = ["full"] }
-lemmy_db_views_reports = { workspace = true, features = ["full"] }
 lemmy_db_views_local_image = { workspace = true, features = ["full"] }
 lemmy_db_views_inbox_combined = { workspace = true, features = ["full"] }
 lemmy_db_views_modlog_combined = { workspace = true, features = ["full"] }
