+++ crates/api/api/src/local_user/report_count.rs
@@ -1,8 +1,10 @@
 use actix_web::web::{Data, Json, Query};
 use lemmy_api_utils::{context::LemmyContext, utils::check_community_mod_of_any_or_admin_action};
 use lemmy_db_views_local_user::LocalUserView;
-use lemmy_db_views_report_combined::ReportCombinedViewInternal;
-use lemmy_db_views_reports::api::{GetReportCount, GetReportCountResponse};
+use lemmy_db_views_report_combined::{
+  api::{GetReportCount, GetReportCountResponse},
+  ReportCombinedViewInternal,
+};
 use lemmy_utils::error::LemmyResult;
 
 pub async fn report_count(
