+++ crates/api/api/src/reports/comment_report/create.rs
@@ -13,9 +13,10 @@ use lemmy_db_schema::{
 };
 use lemmy_db_views_comment::CommentView;
 use lemmy_db_views_local_user::LocalUserView;
-use lemmy_db_views_reports::{
+use lemmy_db_views_report_combined::{
   api::{CommentReportResponse, CreateCommentReport},
   CommentReportView,
+  ReportCombinedViewInternal,
 };
 use lemmy_db_views_site::SiteView;
 use lemmy_email::admin::send_new_report_email_to_admins;
@@ -63,7 +64,8 @@ pub async fn create_comment_report(
   let report = CommentReport::report(&mut context.pool(), &report_form).await?;
 
   let comment_report_view =
-    CommentReportView::read(&mut context.pool(), report.id, person_id).await?;
+    ReportCombinedViewInternal::read_comment_report(&mut context.pool(), report.id, person_id)
+      .await?;
 
   // Email the admins
   let local_site = SiteView::read_local(&mut context.pool()).await?.local_site;
