+++ crates/api/api/src/reports/post_report/resolve.rs
@@ -8,9 +8,10 @@ use lemmy_api_utils::{
 };
 use lemmy_db_schema::{source::post_report::PostReport, traits::Reportable};
 use lemmy_db_views_local_user::LocalUserView;
-use lemmy_db_views_reports::{
+use lemmy_db_views_report_combined::{
   api::{PostReportResponse, ResolvePostReport},
   PostReportView,
+  ReportCombinedViewInternal,
 };
 use lemmy_utils::error::LemmyResult;
 
@@ -22,7 +23,8 @@ pub async fn resolve_post_report(
 ) -> LemmyResult<Json<PostReportResponse>> {
   let report_id = data.report_id;
   let person_id = local_user_view.person.id;
-  let report = PostReportView::read(&mut context.pool(), report_id, person_id).await?;
+  let report =
+    ReportCombinedViewInternal::read_post_report(&mut context.pool(), report_id, person_id).await?;
 
   let person_id = local_user_view.person.id;
   check_community_mod_action(
@@ -39,7 +41,8 @@ pub async fn resolve_post_report(
     PostReport::unresolve(&mut context.pool(), report_id, person_id).await?;
   }
 
-  let post_report_view = PostReportView::read(&mut context.pool(), report_id, person_id).await?;
+  let post_report_view =
+    ReportCombinedViewInternal::read_post_report(&mut context.pool(), report_id, person_id).await?;
 
   ActivityChannel::submit_activity(
     SendActivityData::SendResolveReport {
