+++ crates/api/api/src/reports/private_message_report/create.rs
@@ -9,11 +9,11 @@ use lemmy_db_schema::{
   traits::{Crud, Reportable},
 };
 use lemmy_db_views_local_user::LocalUserView;
-use lemmy_db_views_reports::{
+use lemmy_db_views_report_combined::{
   api::{CreatePrivateMessageReport, PrivateMessageReportResponse},
   PrivateMessageReportView,
 };
-use lemmy_db_views_site::SiteView;
+use lemmy_db_views_site::SiteView;use lemmy_db_views_report_combined::ReportCombinedViewInternal;
 use lemmy_email::admin::send_new_report_email_to_admins;
 use lemmy_utils::error::{LemmyErrorType, LemmyResult};
 
@@ -45,7 +45,7 @@ pub async fn create_pm_report(
   let report = PrivateMessageReport::report(&mut context.pool(), &report_form).await?;
 
   let private_message_report_view =
-    PrivateMessageReportView::read(&mut context.pool(), report.id).await?;
+  ReportCombinedViewInternal::read_private_message_report(&mut context.pool(), report.id, person_id).await?;
 
   // Email the admins
   let local_site = SiteView::read_local(&mut context.pool()).await?.local_site;
