+++ crates/api/api/src/reports/private_message_report/resolve.rs
@@ -2,9 +2,10 @@ use actix_web::web::{Data, Json};
 use lemmy_api_utils::{context::LemmyContext, utils::is_admin};
 use lemmy_db_schema::{source::private_message_report::PrivateMessageReport, traits::Reportable};
 use lemmy_db_views_local_user::LocalUserView;
-use lemmy_db_views_reports::{
+use lemmy_db_views_report_combined::{
   api::{PrivateMessageReportResponse, ResolvePrivateMessageReport},
   PrivateMessageReportView,
+  ReportCombinedViewInternal,
 };
 use lemmy_utils::error::LemmyResult;
 
@@ -23,8 +24,12 @@ pub async fn resolve_pm_report(
     PrivateMessageReport::unresolve(&mut context.pool(), report_id, person_id).await?;
   }
 
-  let private_message_report_view =
-    PrivateMessageReportView::read(&mut context.pool(), report_id).await?;
+  let private_message_report_view = ReportCombinedViewInternal::read_private_message_report(
+    &mut context.pool(),
+    report_id,
+    person_id,
+  )
+  .await?;
 
   Ok(Json(PrivateMessageReportResponse {
     private_message_report_view,
