+++ crates/db_views/report_combined/Cargo.toml
@@ -24,10 +24,9 @@ full = [
   "lemmy_db_schema/full",
   "lemmy_db_schema_file/full",
 ]
-ts-rs = ["dep:ts-rs", "lemmy_db_schema/ts-rs", "lemmy_db_views_reports/ts-rs"]
+ts-rs = ["dep:ts-rs", "lemmy_db_schema/ts-rs"]
 
 [dependencies]
-lemmy_db_views_reports = { workspace = true }
 lemmy_db_views_local_user = { workspace = true }
 lemmy_db_schema = { workspace = true }
 lemmy_utils = { workspace = true, optional = true }
