+++ crates/db_views/report_combined/src/lib.rs
@@ -10,18 +10,14 @@ use lemmy_db_schema::source::{
   private_message::PrivateMessage,
   private_message_report::PrivateMessageReport,
 };
-use lemmy_db_views_reports::{
-  CommentReportView,
-  CommunityReportView,
-  PostReportView,
-  PrivateMessageReportView,
-};
 use serde::{Deserialize, Serialize};
+use serde_with::skip_serializing_none;
 #[cfg(feature = "full")]
 use {
   diesel::{dsl::Nullable, NullableExpressionMethods, Queryable, Selectable},
+  lemmy_db_schema::utils::queries::{local_user_is_admin, person1_select, person2_select},
   lemmy_db_schema::{
-    utils::queries::{local_user_is_admin, person1_select, person2_select},
+    utils::queries::{comment_creator_is_admin, post_creator_is_admin},
     Person1AliasAllColumnsTuple,
     Person2AliasAllColumnsTuple,
   },
@@ -97,4 +93,147 @@ pub enum ReportCombinedView {
   Comment(CommentReportView),
   PrivateMessage(PrivateMessageReportView),
   Community(CommunityReportView),
+}
+
+#[skip_serializing_none]
+#[derive(Debug, PartialEq, Eq, Serialize, Deserialize, Clone)]
+#[cfg_attr(feature = "full", derive(Queryable, Selectable))]
+#[cfg_attr(feature = "full", diesel(check_for_backend(diesel::pg::Pg)))]
+#[cfg_attr(feature = "ts-rs", derive(ts_rs::TS))]
+#[cfg_attr(feature = "ts-rs", ts(optional_fields, export))]
+/// A private message report view.
+pub struct PrivateMessageReportView {
+  #[cfg_attr(feature = "full", diesel(embed))]
+  pub private_message_report: PrivateMessageReport,
+  #[cfg_attr(feature = "full", diesel(embed))]
+  pub private_message: PrivateMessage,
+  #[cfg_attr(feature = "full", diesel(embed))]
+  pub creator: Person,
+  #[cfg_attr(feature = "full",
+    diesel(
+      select_expression_type = Person1AliasAllColumnsTuple,
+      select_expression = person1_select()
+    )
+  )]
+  pub private_message_creator: Person,
+  #[cfg_attr(feature = "full",
+    diesel(
+      select_expression_type = Nullable<Person2AliasAllColumnsTuple>,
+      select_expression = person2_select().nullable()
+    )
+  )]
+  pub resolver: Option<Person>,
+}
+
+#[skip_serializing_none]
+#[derive(Debug, PartialEq, Serialize, Deserialize, Clone)]
+#[cfg_attr(feature = "full", derive(Queryable, Selectable))]
+#[cfg_attr(feature = "full", diesel(check_for_backend(diesel::pg::Pg)))]
+#[cfg_attr(feature = "ts-rs", derive(ts_rs::TS))]
+#[cfg_attr(feature = "ts-rs", ts(optional_fields, export))]
+/// A comment report view.
+pub struct CommentReportView {
+  #[cfg_attr(feature = "full", diesel(embed))]
+  pub comment_report: CommentReport,
+  #[cfg_attr(feature = "full", diesel(embed))]
+  pub comment: Comment,
+  #[cfg_attr(feature = "full", diesel(embed))]
+  pub post: Post,
+  #[cfg_attr(feature = "full", diesel(embed))]
+  pub community: Community,
+  #[cfg_attr(feature = "full", diesel(embed))]
+  pub creator: Person,
+  #[cfg_attr(feature = "full",
+    diesel(
+      select_expression_type = Person1AliasAllColumnsTuple,
+      select_expression = person1_select()
+    )
+  )]
+  pub comment_creator: Person,
+  #[cfg_attr(feature = "full", diesel(embed))]
+  pub comment_actions: Option<CommentActions>,
+  #[cfg_attr(feature = "full",
+    diesel(
+      select_expression_type = Nullable<Person2AliasAllColumnsTuple>,
+      select_expression = person2_select().nullable()
+    )
+  )]
+  pub resolver: Option<Person>,
+  #[cfg_attr(feature = "full", diesel(embed))]
+  pub person_actions: Option<PersonActions>,
+  #[cfg_attr(feature = "full", diesel(embed))]
+  pub community_actions: Option<CommunityActions>,
+  #[cfg_attr(feature = "full",
+    diesel(
+      select_expression = comment_creator_is_admin()
+    )
+  )]
+  pub creator_is_admin: bool,
+}
+
+#[skip_serializing_none]
+#[derive(Debug, PartialEq, Serialize, Deserialize, Clone)]
+#[cfg_attr(feature = "full", derive(Queryable, Selectable))]
+#[cfg_attr(feature = "full", diesel(check_for_backend(diesel::pg::Pg)))]
+#[cfg_attr(feature = "ts-rs", derive(ts_rs::TS))]
+#[cfg_attr(feature = "ts-rs", ts(optional_fields, export))]
+/// A community report view.
+pub struct CommunityReportView {
+  #[cfg_attr(feature = "full", diesel(embed))]
+  pub community_report: CommunityReport,
+  #[cfg_attr(feature = "full", diesel(embed))]
+  pub community: Community,
+  #[cfg_attr(feature = "full", diesel(embed))]
+  pub creator: Person,
+  #[cfg_attr(feature = "full",
+    diesel(
+      select_expression_type = Nullable<Person2AliasAllColumnsTuple>,
+      select_expression = person2_select().nullable()
+    )
+  )]
+  pub resolver: Option<Person>,
+}
+
+#[skip_serializing_none]
+#[derive(Debug, PartialEq, Serialize, Deserialize, Clone)]
+#[cfg_attr(feature = "full", derive(Queryable, Selectable))]
+#[cfg_attr(feature = "full", diesel(check_for_backend(diesel::pg::Pg)))]
+#[cfg_attr(feature = "ts-rs", derive(ts_rs::TS))]
+#[cfg_attr(feature = "ts-rs", ts(optional_fields, export))]
+/// A post report view.
+pub struct PostReportView {
+  #[cfg_attr(feature = "full", diesel(embed))]
+  pub post_report: PostReport,
+  #[cfg_attr(feature = "full", diesel(embed))]
+  pub post: Post,
+  #[cfg_attr(feature = "full", diesel(embed))]
+  pub community: Community,
+  #[cfg_attr(feature = "full", diesel(embed))]
+  pub creator: Person,
+  #[cfg_attr(feature = "full",
+    diesel(
+      select_expression_type = Person1AliasAllColumnsTuple,
+      select_expression = person1_select()
+    )
+  )]
+  pub post_creator: Person,
+  #[cfg_attr(feature = "full", diesel(embed))]
+  pub community_actions: Option<CommunityActions>,
+  #[cfg_attr(feature = "full", diesel(embed))]
+  pub post_actions: Option<PostActions>,
+  #[cfg_attr(feature = "full", diesel(embed))]
+  pub person_actions: Option<PersonActions>,
+  #[cfg_attr(feature = "full",
+    diesel(
+      select_expression_type = Nullable<Person2AliasAllColumnsTuple>,
+      select_expression = person2_select().nullable()
+    )
+  )]
+  pub resolver: Option<Person>,
+  #[cfg_attr(feature = "full",
+    diesel(
+      select_expression = post_creator_is_admin()
+    )
+  )]
+  pub creator_is_admin: bool,
 }
