--- crates/db_views/reports/Cargo.toml
@@ -1,36 +1,0 @@
-[package]
-name = "lemmy_db_views_reports"
-version.workspace = true
-edition.workspace = true
-description.workspace = true
-license.workspace = true
-homepage.workspace = true
-documentation.workspace = true
-repository.workspace = true
-rust-version.workspace = true
-
-[lib]
-doctest = false
-
-[lints]
-workspace = true
-
-[features]
-full = [
-  "lemmy_utils",
-  "diesel",
-  "diesel-async",
-  "lemmy_db_schema/full",
-  "lemmy_db_schema_file/full",
-]
-ts-rs = ["dep:ts-rs", "lemmy_db_schema/ts-rs"]
-
-[dependencies]
-lemmy_db_schema = { workspace = true }
-lemmy_utils = { workspace = true, optional = true }
-lemmy_db_schema_file = { workspace = true }
-diesel = { workspace = true, optional = true }
-diesel-async = { workspace = true, optional = true }
-serde = { workspace = true }
-serde_with = { workspace = true }
-ts-rs = { workspace = true, optional = true }
